Add basic prediction based backtesting strategy
Introduce simple backtesting strategy that manages positions based on
external predictions. This allows us to decouple the prediction logic
from the general backtesting strategy.

The logic of the prediction-based strategy is as follows:
- On each feed update, recompute the latest prediction
- If the prediction is more than half a tick (default $0.01) above or
below the current price then submit an order to buy or sell accordingly
- The order size is configurable (default 100) and capped by a max
position size (default 10000), as well as the outstanding account cash
balance
- On the start of the next feed update, cancel previous active orders
that have not been filled