Proof of concept wiring backtesting into validation
- Add support for including feed for additional input instruments in
backtesting in addition to target instrument.
- Add new predictor implementation that delegates to the encoder/decoder
model.
- Add partially stubbed out logic to use backtesting pnl for model
validation.
- Add example "test1" constituent data to use as additional instrument
for testing purposes.

Still need to add logic to track the relevant symbols and dates in the
da rnn model to pass valid inputs into the backtesting. Additionally, we
may want to make sure that the training and validation outputs are more
comparable. Currently, they are training loss vs validation pnl.